Updated stat function to yield more info Allow changing of channels and channel layout Added subtitles to sample video Added test cases

+++ php_av.h
@@ -163,14 +163,14 @@ int av_optimize_mov_file(AVIOContext *pb);
 int av_get_element_double(zval *array, const char *key, double *p_value);
 int av_get_element_long(zval *array, const char *key, long *p_value);
 int av_get_element_string(zval *array, const char *key, char **p_value);
-int av_get_element_stringl(zval *array, const char *key, char **p_value, uint32_t *p_len);
+int av_get_element_stringl(zval *array, const char *key, char **p_value, long *p_len);
 int av_get_element_hash(zval *array, const char *key, HashTable **p_hash);
 int av_get_element_resource(zval *array, const char *key, zval **p_res);
 
 void av_set_element_long(zval *array, const char *key, long value);
 void av_set_element_double(zval *array, const char *key, double value);
 void av_set_element_string(zval *array, const char *key, const char *value);
-void av_set_element_stringl(zval *array, const char *key, const char *value, uint32_t value_length);
+void av_set_element_stringl(zval *array, const char *key, const char *value, long value_length);
 
 zval *av_create_gd_image(uint32_t width, uint32_t height TSRMLS_DC);
